flat node linked list cleanup, move edit.jsx
